Remove dangling underscores from range properties
Originally, these were to be implemented using property getters/setters,
where modifying the value updated the instance's display. This was later
deemed overkill, since the likelihood of developers needing to customise
the range lengths post-construction didn't justify the extra code.

If the relevance of that paragraph isn't obvious to you, underscores are
used to identify variables that hold the internal value of an externally
accessible method or property.